# doc/api.yaml
...
openapi: 3.0.3
info:
    title: 'WASAText API'
    description: |-
        A modern chat application API that allows users to create groups,
        start conversations, send messages, react to messages, and manage
        contacts.

        Features include:
        - Real-time messaging with WebSocket support
        - Image messaging with automatic compression
        - Message reactions with toggle behavior (prevents duplicates)
        - Group chat management
        - Contact management
        - Message forwarding (including images)
        - User profiles with photos
    version: '1.2.0'
security:
    - bearerAuth: []

# Note: All authenticated endpoints (those requiring bearerAuth) will return 
# a 401 Unauthorized response if authentication fails. See the UnauthorizedError 
# response definition in components for details.
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: UUID
            description: |-
                Bearer token authentication using User UUID as the token.
                The token should be the UUID of an existing user in the system.
    schemas:
        User:
            type: object
            properties:
                id:
                    type: string
                    example: 'f2555a8a-2e66-4326-9588-20e7e298d615'
                    description: UUID of the user
                username:
                    type: string
                    example: 'Alice'
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 16
                picture:
                    type: string
                    example: 'https://example.com/avatar.jpg'
                    description: URL to user's profile picture
            required:
                - id
                - username
...
        Message:
            type: object
            properties:
                id:
                    type: string
                    example: '85f9cc23-48d7-448d-9172-7d5fd9cd4858'
                    description: UUID of the message
                senderId:
                    type: string
                    example: 'f2555a8a-2e66-4326-9588-20e7e298d615'
                    description: UUID of the message sender
                text:
                    type: string
                    example: 'Hello, how are you?'
                    description: Content of the message
                imageUrl:
                    type: string
                    example: 'data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQ...'
                    description: Base64 encoded image data with MIME type prefix (optional)
                senderUsername:
                    type: string
                    example: 'Alice'
                    description: Username of the message sender
                time:
                    type: string
                    example: '2025-01-01T12:00:00Z'
                    description: Timestamp when the message was sent
                reactions:
                    type: object
                    additionalProperties:
                        type: object
                        properties:
                            count:
                                type: integer
                                example: 3
                            users:
                                type: array
                                items:
                                    type: string
                                example: ['Alice', 'Bob']
                    description: >-
                        Reactions to the message (emoji -> reaction data)
            required:
                - id
                - senderId
                - senderUsername
...
        Contact:
            type: object
            properties:
                id:
                    type: string
                    example: 'contact-uuid'
                    description: UUID of the contact
                userId:
                    type: string
                    example: 'user-uuid'
                    description: UUID of the user who owns this contact
                contactUserId:
                    type: string
                    example: 'contact-user-uuid'
                    description: UUID of the user being contacted
                username:
                    type: string
                    example: 'Bob'
                    description: Username of the contact
            required:
                - id
                - userId
                - contactUserId
                - username
        LoginRequest:
            type: object
            properties:
                name:
                    type: string
                    example: 'Alice'
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 16
            required:
                - name
        LoginResponse:
            type: object
            properties:
                identifier:
                    type: string
                    example: 'f2555a8a-2e66-4326-9588-20e7e298d615'
                    description: UUID of the logged in user
            required:
                - identifier
        SendMessageRequest:
            type: object
            properties:
                content:
                    type: string
                    example: 'Hello everyone!'
                    description: Content of the message to send
                imageUrl:
                    type: string
                    example: 'data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQ...'
                    description: Base64 encoded image data with MIME type prefix (optional)
            anyOf:
                - required: [content]
                - required: [imageUrl]
...
        AddContactRequest:
            type: object
            properties:
                contactUserId:
                    type: string
                    example: 'user-uuid'
                    description: UUID of the user to add as contact
            required:
                - contactUserId
...
        Error:
            type: object
            properties:
                message:
                    type: string
                    example: 'Invalid input provided'
                    description: Error message describing what went wrong
            required:
                - message
    responses:
        UnauthorizedError:
            description: Authentication information is missing or invalid
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
                    examples:
                        missing_header:
                            summary: Missing Authorization header
                            value:
                                message: "Authorization header required"
                        invalid_format:
                            summary: Invalid authorization format
                            value:
                                message: "Invalid authorization format. Use 'Bearer <token>'"
                        invalid_token:
                            summary: Invalid token format
                            value:
                                message: "Invalid token format"
                        user_not_found:
                            summary: User not found
                            value:
                                message: "Invalid token - user not found"
paths:
    /session:
        post:
            tags: ['Authentication']
            summary: User login
            description: |-
                If the user does not exist, it will be created and an
                identifier is returned.
                If the user exists, the user identifier is returned.
            operationId: doLogin
            requestBody:
                description: User login details
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
                required: true
            responses:
                '201':
                    description: User login successful
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
                '400':
                    description: Invalid input - name must be 3-16 characters
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /users:
        get:
            tags: ['Users']
            summary: List all users
            description: Returns a list of all users in the system
            operationId: listUsers
            responses:
                '200':
                    description: List of users
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/User'
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '500':
                    description: Internal server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /users/{id}:
        parameters:
            - name: id
              in: path
              description: UUID of the user
              required: true
              schema:
                  type: string
                  format: uuid
        put:
            tags: ['Users']
            summary: Update username
            description: Set or update the user's username
            operationId: setMyUserName
            requestBody:
                description: New username as a JSON string
                content:
                    application/json:
                        schema:
                            type: string
                            minLength: 3
                            maxLength: 16
                            example: 'Alice'
                required: true
            responses:
                '200':
                    description: Username updated successfully
                '400':
                    description: Invalid username format
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '404':
                    description: User not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...
    /users/{id}/conversations/{conversationId}/messages:
        parameters:
            - name: id
              in: path
              description: UUID of the user
              required: true
              schema:
                  type: string
                  format: uuid
            - name: conversationId
              in: path
              description: UUID of the conversation
              required: true
              schema:
                  type: string
                  format: uuid
        get:
            tags: ['Messages']
            summary: Get messages
            description: Get all messages from a conversation
            operationId: getMessages
            responses:
                '200':
                    description: List of messages
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Message'
                '404':
                    description: Conversation not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        post:
            tags: ['Messages']
            summary: Send message
            description: Send a new text message or image to the conversation. Images should be provided as base64-encoded data URLs.
            operationId: sendMessage
            requestBody:
                description: Message content (text and/or image)
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SendMessageRequest'
                required: true
            responses:
                '201':
                    description: Message sent successfully
                    content:
                        application/json:
                            schema:
                                type: string
                                description: UUID of the created message
                '400':
                    description: Invalid message content
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Conversation not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...
    /users/{id}/contacts:
        parameters:
            - name: id
              in: path
              description: UUID of the user
              required: true
              schema:
                  type: string
                  format: uuid
        get:
            tags: ['Contacts']
            summary: List contacts
            description: Get all contacts for the user
            operationId: listContacts
            responses:
                '200':
                    description: List of contacts
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Contact'
                '404':
                    description: User not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        post:
            tags: ['Contacts']
            summary: Add contact
            description: Add a new contact for the user
            operationId: addContact
            requestBody:
                description: Contact details
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AddContactRequest'
                required: true
            responses:
                '201':
                    description: Contact added successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Contact'
                '400':
                    description: Invalid contact data
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: User not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /users/{id}/contacts/{contactId}:
        parameters:
            - name: id
              in: path
              description: UUID of the user
              required: true
              schema:
                  type: string
                  format: uuid
            - name: contactId
              in: path
              description: UUID of the contact
              required: true
              schema:
                  type: string
                  format: uuid
        delete:
            tags: ['Contacts']
            summary: Remove contact
            description: Remove a contact from the user's contact list
            operationId: removeContact
            responses:
                '200':
                    description: Contact removed successfully
                '404':
                    description: Contact not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...